Add support for Tailwind CSS 2.0
This change adds support for Tailwind CSS v2.0. Before this change, your package manager of choice would complain about an unmet peer dependency if it was used it with Tailswind CSS 2 or newer.

The blog post and the upgrade guide do not mention any changes for plugins, I also verified the functionality in a project of mine. Everything still works as expected as far as I can see. :)

https://blog.tailwindcss.com/tailwindcss-v2
https://tailwindcss.com/docs/upgrading-to-v2